Move responses code to its own file.
Make it's a bit easier to understand.
