+++ js/src/wrappers/oai.ts
@@ -9,7 +9,7 @@ import {
 import { getCurrentUnixTimestamp, isEmpty, filterFrom } from "../util";
 import { mergeDicts } from "@braintrust/core";
 import { proxyCreate, TimedSpan } from "./wrapper_utils";
-
+import { responsesProxy } from "./oai_responses";
 interface BetaLike {
   chat: {
     completions: {
@@ -126,163 +126,6 @@ export function wrapOpenAIv4<T extends OpenAILike>(openai: T): T {
         return betaProxy;
       }
       return Reflect.get(target, name, receiver);
-    },
-  });
-}
-
-function responsesProxy(openai: OpenAILike) {
-  // This was added in v4.87.0 of the openai-node library
-  if (!openai.responses) {
-    return openai;
-  }
-
-  return new Proxy(openai.responses, {
-    get(target, name, receiver) {
-      if (name === "create") {
-        return responsesCreateProxy(target.create.bind(target));
-      } else if (name === "stream") {
-        return responsesStreamProxy(target.stream.bind(target));
-      }
-      return Reflect.get(target, name, receiver);
-    },
-  });
-}
-
-function responsesCreateProxy(target: any): (params: any) => Promise<any> {
-  const hooks = {
-    name: "openai.responses",
-    toSpanFunc: parseSpanFromResponseCreateParams,
-    resultToEventFunc: parseEventFromResponseCreateResult,
-    traceStreamFunc: traceResponseCreateStream,
-  };
-
-  return proxyCreate(target, hooks);
-}
-
-// convert response.create params into a span
-function parseSpanFromResponseCreateParams(params: any): TimedSpan {
-  // responses.create is meant to take a single message and instruction.
-  // Convert that to the form our backend expects.
-  let input = params.input;
-  if (params.instructions) {
-    input = [
-      { role: "user", content: input },
-      { role: "system", content: params.instructions },
-    ];
-  }
-
-  const spanArgs = {
-    name: "openai.responses.create",
-    spanAttributes: {
-      type: SpanTypeAttribute.LLM,
-      provider: "openai",
-    },
-    event: {
-      input,
-      metadata: filterFrom(params, ["input", "instructions"]),
-    },
-    startTime: getCurrentUnixTimestamp(),
-  };
-  return {
-    span: startSpan(spanArgs),
-    start: spanArgs.startTime,
-  };
-}
-
-// convert response.create result into an event
-function parseEventFromResponseCreateResult(result: any) {
-  return {
-    output: result?.output_text || "",
-    metrics: parseMetricsFromUsage(result?.usage),
-  };
-}
-
-function traceResponseCreateStream(
-  stream: AsyncIterator<any>,
-  timedSpan: TimedSpan,
-) {
-  const span = timedSpan.span;
-  let ttft = -1;
-  return async function <T>(...args: [any]): Promise<IteratorResult<T>> {
-    const result = await stream.next(...args);
-
-    if (ttft === -1) {
-      ttft = getCurrentUnixTimestamp() - timedSpan.start;
-      span.log({ metrics: { time_to_first_token: ttft } });
-    }
-
-    if (result.done) {
-      span.end();
-      return result;
-    }
-
-    const item = result.value;
-    if (!item || !item?.type || !item?.response) {
-      return result; // unexpected
-    }
-
-    const event = parseLogFromItem(item);
-    span.log(event);
-
-    return result;
-  };
-}
-
-function parseLogFromItem(item: any): {} {
-  if (!item || !item?.type || !item?.response) {
-    return {};
-  }
-
-  const response = item.response;
-
-  switch (item.type) {
-    case "response.completed":
-      // I think there is usually only one output, but since they are arrays
-      // we'll collect them all just in case.
-      const texts = [];
-      for (const output of response?.output || []) {
-        for (const content of output?.content || []) {
-          if (content?.type === "output_text") {
-            texts.push(content.text);
-          }
-        }
-      }
-      return {
-        output: texts.join(""),
-        metrics: parseMetricsFromUsage(response?.usage),
-      };
-    default:
-      return {};
-  }
-}
-
-function responsesStreamProxy(target: any): (params: any) => Promise<any> {
-  return new Proxy(target, {
-    apply(target, thisArg, argArray) {
-      const responseStream: any = Reflect.apply(target, thisArg, argArray);
-      if (!argArray || argArray.length === 0) {
-        return responseStream;
-      }
-
-      const timedSpan = parseSpanFromResponseCreateParams(argArray[0]);
-      const span = timedSpan.span;
-
-      let ttft = -1;
-
-      responseStream.on("event", (event: any) => {
-        if (ttft === -1) {
-          ttft = getCurrentUnixTimestamp() - timedSpan.start;
-          span.log({ metrics: { time_to_first_token: ttft } });
-        }
-        const logEvent = parseLogFromItem(event);
-        span.log(logEvent);
-      });
-
-      responseStream.on("end", () => {
-        span.end();
-      });
-
-      return responseStream;
     },
   });
 }
@@ -749,52 +592,4 @@ class WrapperStream<Item> implements AsyncIterable<Item> {
       this.span.end();
     }
   }
-}
-
-type StartedSpan = {
-  span: Span;
-  startTime: number;
-};
-
-function parseMetricsFromUsage(usage: any): Record<string, number> {
-  if (!usage) {
-    return {};
-  }
-
-  // example : {
-  //   input_tokens: 14,
-  //   input_tokens_details: { cached_tokens: 0 },
-  //   output_tokens: 8,
-  //   output_tokens_details: { reasoning_tokens: 0 },
-  //   total_tokens: 22
-  // }
-
-  const keys = [
-    ["input_tokens", "prompt_tokens"],
-    ["output_tokens", "completion_tokens"],
-    ["total_tokens", "tokens"],
-  ];
-
-  const metrics: Record<string, number> = {};
-  for (const [src, target] of keys) {
-    const value = usage[src];
-    if (value !== undefined && value !== null) {
-      metrics[target] = value;
-    }
-  }
-
-  const details: string[] = ["input", "output"];
-  for (const src of details) {
-    const details = usage[`${src}_tokens_details`];
-    if (details) {
-      for (const [key, value] of Object.entries(details)) {
-        const metricName = `${src}_${key}` as string;
-        if (typeof value === "number") {
-          metrics[metricName] = value;
-        }
-      }
-    }
-  }
-
-  return metrics;
 }
